# Pass around a DataView instead of individual kind/payload fields