--- a/Sources/JavaScriptKit/XcodeSupport.swift
+++ b/Sources/JavaScriptKit/XcodeSupport.swift
@@ -9,34 +9,22 @@ import _CJavaScriptKit
     func _set_prop(
         _: JavaScriptObjectRef,
         _: UnsafePointer<Int8>!, _: Int32,
-        _: JavaScriptValueKind,
-        _: JavaScriptPayload1,
-        _: JavaScriptPayload2,
-        _: JavaScriptPayload3
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _get_prop(
         _: JavaScriptObjectRef,
         _: UnsafePointer<Int8>!, _: Int32,
-        _: UnsafeMutablePointer<JavaScriptValueKind>!,
-        _: UnsafeMutablePointer<JavaScriptPayload1>!,
-        _: UnsafeMutablePointer<JavaScriptPayload2>!,
-        _: UnsafeMutablePointer<JavaScriptPayload3>!
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _set_subscript(
         _: JavaScriptObjectRef,
         _: Int32,
-        _: JavaScriptValueKind,
-        _: JavaScriptPayload1,
-        _: JavaScriptPayload2,
-        _: JavaScriptPayload3
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _get_subscript(
         _: JavaScriptObjectRef,
         _: Int32,
-        _: UnsafeMutablePointer<JavaScriptValueKind>!,
-        _: UnsafeMutablePointer<JavaScriptPayload1>!,
-        _: UnsafeMutablePointer<JavaScriptPayload2>!,
-        _: UnsafeMutablePointer<JavaScriptPayload3>!
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _load_string(
         _: JavaScriptObjectRef,
@@ -45,19 +33,13 @@ import _CJavaScriptKit
     func _call_function(
         _: JavaScriptObjectRef,
         _: UnsafePointer<RawJSValue>!, _: Int32,
-        _: UnsafeMutablePointer<JavaScriptValueKind>!,
-        _: UnsafeMutablePointer<JavaScriptPayload1>!,
-        _: UnsafeMutablePointer<JavaScriptPayload2>!,
-        _: UnsafeMutablePointer<JavaScriptPayload3>!
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _call_function_with_this(
         _: JavaScriptObjectRef,
         _: JavaScriptObjectRef,
         _: UnsafePointer<RawJSValue>!, _: Int32,
-        _: UnsafeMutablePointer<JavaScriptValueKind>!,
-        _: UnsafeMutablePointer<JavaScriptPayload1>!,
-        _: UnsafeMutablePointer<JavaScriptPayload2>!,
-        _: UnsafeMutablePointer<JavaScriptPayload3>!
+        _: UnsafePointer<RawJSValue>!
     ) { fatalError() }
     func _call_new(
         _: JavaScriptObjectRef,