--- a/Sources/_CJavaScriptKit/include/_CJavaScriptKit.h
+++ b/Sources/_CJavaScriptKit/include/_CJavaScriptKit.h
@@ -25,35 +25,28 @@ typedef struct {
   JavaScriptValueKind kind;
   JavaScriptPayload1 payload1;
   JavaScriptPayload2 payload2;
-  JavaScriptPayload3 payload3;
 } RawJSValue;
 
 #if __wasm32__
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_set_prop"))) extern void
 _set_prop(const JavaScriptObjectRef _this, const char *prop, const int length,
-          const JavaScriptValueKind kind, const JavaScriptPayload1 payload1,
-          const JavaScriptPayload2 payload2, const JavaScriptPayload3 payload3);
+          const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_get_prop"))) extern void
 _get_prop(const JavaScriptObjectRef _this, const char *prop, const int length,
-          JavaScriptValueKind *kind, JavaScriptPayload1 *payload1,
-          JavaScriptPayload2 *payload2, JavaScriptPayload3 *payload3);
+          const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_set_subscript"))) extern void
 _set_subscript(const JavaScriptObjectRef _this, const int length,
-               const JavaScriptValueKind kind,
-               const JavaScriptPayload1 payload1,
-               const JavaScriptPayload2 payload2,
-               const JavaScriptPayload3 payload3);
+               const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_get_subscript"))) extern void
 _get_subscript(const JavaScriptObjectRef _this, const int length,
-               JavaScriptValueKind *kind, JavaScriptPayload1 *payload1,
-               JavaScriptPayload2 *payload2, JavaScriptPayload3 *payload3);
+               const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_load_string"))) extern void
@@ -62,20 +55,14 @@ _load_string(const JavaScriptObjectRef ref, unsigned char *buffer);
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_call_function"))) extern void
 _call_function(const JavaScriptObjectRef ref, const RawJSValue *argv,
-               const int argc, JavaScriptValueKind *result_kind,
-               JavaScriptPayload1 *result_payload1,
-               JavaScriptPayload2 *result_payload2,
-               JavaScriptPayload3 *result_payload3);
+               const int argc, const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_call_function_with_this"))) extern void
 _call_function_with_this(const JavaScriptObjectRef _this,
                          const JavaScriptObjectRef func_ref,
                          const RawJSValue *argv, const int argc,
-                         JavaScriptValueKind *result_kind,
-                         JavaScriptPayload1 *result_payload1,
-                         JavaScriptPayload2 *result_payload2,
-                         JavaScriptPayload3 *result_payload3);
+                         const RawJSValue *rawJSValue);
 
 __attribute__((__import_module__("javascript_kit"),
                __import_name__("swjs_call_new"))) extern void