Also change memory layout for numbers:

There are now only two payload slots. Numbers use them as a single Double, while every other type just uses them as two UInt32s

Originally part of #22.