@kateinoigakukun 2020-08-02
Thanks for your cooperation.

Did you check performance benchmark suite? You can check benchmark by `make benchmark` in `IntegrationTests` directory.

I found some performance regression on number interoperation. 

Before

```
Running 'Serialization/Write JavaScript number directly' ...
done 118.31193709373474 ms
Running 'Serialization/Write JavaScript string directly' ...
done 119.49532198905945 ms
Running 'Serialization/Swift Int to JavaScript' ...
done 3457.1316990852356 ms
Running 'Serialization/Swift String to JavaScript' ...
done 5258.867609977722 ms
Running 'Object heap/Increment and decrement RC' ...
done 3061.0713909864426 ms
```

After
```
Running 'Serialization/Write JavaScript number directly' ...
done 118.58276605606079 ms
Running 'Serialization/Write JavaScript string directly' ...
done 120.5656269788742 ms
Running 'Serialization/Swift Int to JavaScript' ...
done 4948.468513011932 ms
Running 'Serialization/Swift String to JavaScript' ...
done 5090.137539982796 ms
Running 'Object heap/Increment and decrement RC' ...
done 2963.7367210388184 ms
```


I'm still reviewing the changes, so please wait for a while.