@kateinoigakukun 2020-08-02
> I’m not exactly sure why they were 24 bytes long before since kind + payload1 + payload2 = 12 bytes and payload3 = 8 bytes which only add up to 20. 

The additional 4 byte is for `double` which is aligned by 8 bytes in C convention