@kateinoigakukun 2020-08-02
I think the regression is due to the indirection of payloads access.
Why don't you pass the two payloads directly as arguments?