@j-f1 2020-08-02
> I think the regression is due to the indirection of payloads access.

That definitely seems to be the case! I’m not sure why they switched to passing in the `DataView`s but I’d be happy to revert that part

> The additional 4 byte is for `double` which is aligned by 8 bytes in C convention

Ah, that makes sense! Since we’re not using C doubles anymore (we’re just encoding them as two ints), this shouldn’t be an issue, right?