@kateinoigakukun 2020-08-02
> That definitely seems to be the case! I’m not sure why they switched to passing in the DataViews but I’d be happy to revert that part

Thanks ❤️ 

> Ah, that makes sense! Since we’re not using C doubles anymore (we’re just encoding them as two ints), this shouldn’t be an issue, right?

Yes, I think it's ok.