@j-f1 2020-08-02
Ok, so I did a test where I removed all the `DataView` stuff (except when decoding numbers (where we need to specify endianness, I think?)) but keeping the new function style where we only pass a single pointer representing the entire JSValue rather than `kind`/`payload1`/`payload2` and it’s somehow slower than all the other styles?

| Test | `master` | single pointer | both |
|-|-|-|-|
| Serialization/Write JavaScript number directly | 123 | 126 | 122 |
| Serialization/Write JavaScript string directly | 130 | 134 | 126 |
| Serialization/Swift Int to JavaScript | 3150 | 5075 | 5083 |
| Serialization/Swift String to JavaScript | 4842 | 4925 | 4838 |
| Object heap/Increment and decrement RC | 2898 | 2817 | 2904 |