@kateinoigakukun 2020-08-02
Hmm, the benchmark results show that decoding number from two 32bit payloads is slower than passing 20 byte payloads directly.

I think we need to optimize the payloads, but this approach seems a little difficult.